// apps/web/src/api/generated/model/modelListDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Open WebUI Klon API
 * OpenAPI spec version: 0.0.0
 */
import type { ModelDto } from './modelDto.ts';
import type { UnavailableConnectionDto } from './unavailableConnectionDto.ts';

export interface ModelListDto {
  models: ModelDto[];
  unavailableConnections: UnavailableConnectionDto[];
}